+++ src/lib.rs
@@ -3,10 +3,10 @@
 #![allow(non_camel_case_types)]
 #![allow(non_snake_case)]
 
-#[cfg(feature = "use-bindgen")]
+#[cfg(any(feature = "use-bindgen", not(docs_rs)))]
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
 
-#[cfg(not(feature = "use-bindgen"))]
+#[cfg(any(not(feature = "use-bindgen"), docs_rs))]
 mod bindings;
-#[cfg(not(feature = "use-bindgen"))]
+#[cfg(any(not(feature = "use-bindgen"), docs_rs))]
 pub use bindings::*;